+++ src/search.ts
@@ -4,6 +4,7 @@ import filterIds from "./search-filters/ids";
 import filterColorIdentity from "./search-filters/color-identity";
 import filterComboData from "./search-filters/combo-data";
 import filterSize from "./search-filters/size";
+import filterTags from "./search-filters/tags";
 import createMessage from "./parse-query/create-message";
 
 import type { SearchResults } from "./types";
@@ -18,6 +19,7 @@ export default async function search(query = ""): Promise<SearchResults> {
   combos = filterColorIdentity(combos, searchParams);
   combos = filterComboData(combos, searchParams);
   combos = filterSize(combos, searchParams);
+  combos = filterTags(combos, searchParams);
 
   return {
     errors,