+++ src/spellbook-api.ts
@@ -43,8 +43,8 @@ function formatApiResponse(
       const prerequisites = SpellbookList.create(combo[12]);
       const steps = SpellbookList.create(combo[13]);
       const results = SpellbookList.create(combo[14]);
-      const hasBannedCard = combo[15] === "true";
-      const hasSpoiledCard = combo[16] === "true";
+      const hasBannedCard = combo[15].toLowerCase() === "true";
+      const hasSpoiledCard = combo[16].toLowerCase() === "true";
 
       return {
         commanderSpellbookId: id,